Jm-updates-tests-mongo (#45)
* fix typo

* adds db

* swap destination for cruise

* move tests to their own file

* fix typo

* updates pages for cruise/cruise-id
